Allow empty paths in syscall wrappers like wstat()
When we resolve an empty path against a working
directory, we return another empty path, so that
we don't accidentally reference the working directory.
(The kernel does essentially the same thing.) If an empty
path is passed to e.g. stat(), the kernel replies with ENOENT.
So let's allow empty paths in wstat(), etc.